refactor: break admin-review-actions.ts into modular components
- Create types/admin-types.ts for TypeScript interfaces and type definitions
- Create utils/wordpress-utils.ts for WordPress-specific functionality
- Create modals/loading-modal.ts for loading modal functionality
- Create modals/prompt-modal.ts for prompt selection modal functionality
- Create services/ajax-handler.ts for AJAX request handling
- Refactor main admin-review-actions.ts to orchestrate modules
- Reduce main file from 283 lines to 79 lines (72% reduction)
- Improve code maintainability and testability through separation of concerns
