+++ src/types/admin-types.ts
@@ -1,0 +1,27 @@
+/**
+ * TypeScript type definitions for admin review actions.
+ *
+ * @since 1.0.0
+ */
+
+export interface Template {
+	value: string;
+	label: string;
+}
+
+export interface WcAiReviewResponder {
+	ajaxurl: string;
+	templates: Template[];
+}
+
+export interface AiResponseData {
+	success: boolean;
+	data: {
+		reply?: string;
+		message?: string;
+	};
+}
+
+declare global {
+	const wcAiReviewResponder: WcAiReviewResponder;
+}
